Move computer window to a separate class
