+++ src/computer.h
@@ -6,9 +6,7 @@
 #include <vector>
 
 #include "calendar.h"
-#include "cursesdef.h"
 
-class player;
 class JsonObject;
 
 // Don't change those! They must stay in this specific order!
@@ -118,14 +116,6 @@ class computer
         void add_failure( const computer_failure &failure );
         void add_failure( computer_failure_type failure );
         void set_access_denied_msg( const std::string &new_msg );
-        // Basic usage
-        /** Shutdown (free w_terminal, etc.) */
-        void shutdown_terminal();
-        /** Handles player use of a computer */
-        void use();
-        /** Returns true if the player successfully hacks the computer. Security = -1 defaults to
-         *  the main system security. */
-        bool hack_attempt( player &p, int Security = -1 );
         // Save/load
         std::string save_data() const;
         void load_data( const std::string &data );
@@ -133,6 +123,7 @@ class computer
         std::string name; // "Jon's Computer", "Lab 6E77-B Terminal Omega"
         int mission_id; // Linked to a mission?
 
+        friend class computer_session;
     private:
         // Difficulty of simply logging in
         int security;
@@ -146,50 +137,10 @@ class computer
         // Can be customized to for example warn the player of potentially lethal
         // consequences like secubots spawning.
         std::string access_denied;
-        // Output window
-        catacurses::window w_terminal;
-        // Pretty border
-        catacurses::window w_border;
         // Misc research notes from json
         static std::vector<std::string> lab_notes;
 
-        // Called by use()
-        void activate_function( computer_action action );
-        // Generally called when we fail a hack attempt
-        void activate_random_failure();
-        // ...but we can also choose a specific failure.
-        void activate_failure( computer_failure_type fail );
-
         void remove_option( computer_action action );
-
-        void mark_refugee_center();
-
-        // OUTPUT/INPUT:
-
-        // Reset to a blank terminal (e.g. at start of usage loop)
-        void reset_terminal();
-        // Prints a line to the terminal (with printf flags)
-        template<typename ...Args>
-        void print_line( const char *text, Args &&... args );
-        // For now, the same as print_line but in red ( TODO: change this?)
-        template<typename ...Args>
-        void print_error( const char *text, Args &&... args );
-        // Wraps and prints a block of text with a 1-space indent
-        template<typename ...Args>
-        void print_text( const char *text, Args &&... args );
-        // Prints code-looking gibberish
-        void print_gibberish_line();
-        // Prints a line and waits for Y/N/Q
-        template<typename ...Args>
-        char query_ynq( const char *text, Args &&... args );
-        // Same as query_ynq, but returns true for y or Y
-        template<typename ...Args>
-        bool query_bool( const char *text, Args &&... args );
-        // Simply wait for any key, returns True
-        template<typename ...Args>
-        bool query_any( const char *text, Args &&... args );
-        // Move the cursor to the beginning of the next line
-        void print_newline();
 };
 
 #endif
